Merge BMC_support and fix kube_proxy (#42)
* Add support for BMC

For BMC, ovs is not needed. Node ls is not needed.

* Add support for BMC

For BMC, ovs is not needed. Node ls is not needed.

* Add support for BMC

For BMC, ovs is not needed. Node ls is not needed.

* Add support for BMC

For BMC, ovs is not needed. Node ls is not needed.
rebased with jumphost support

* Add manual_start_kube_proxy option for playbook

Added a library line in ansible.cfg
Create create_proxy.yaml
run create_proxy

* Refine BMC_spport

No need to install vsphere package if BMC set to false
Add more discription for not-for-BMC parameters

* Update using virtualenv to perform nsx-config

By using virtualenv, the potential pyOpenSSL issue can be solved

* Add openshift version to openshift installation

* Support comma separated MP IPs

* Remove VC components for BMC setup

* Fix for kube_proxy