Removing regiter nsx_config_result
nsx_config_result was causing issue for further tasks to fail.
It was because each time nsx_config_result was overriden and final value was
ok: [fvt1796-246-203.eng.vmware.com] => {
    "msg": {
        "changed": false,
        "skip_reason": "Conditional result was False",
        "skipped": true
    }
}
This doesn't have stdout so caused failer.
Since this is not much userful for user so deleting this for now.